Add VM Ansible playbooks
Before the VMs can be used by GitLab, they need to be prepared.
The first step is using libvirt's lcitool, but then we need some further
setup which is not available with lcitool and so our own Ansible
playbooks are needed - mainly injecting a systemd service which would
notify the host that networking is online, hence the machine being ready
for SSH connections.

Signed-off-by: Erik Skultety <eskultet@redhat.com>