make_base_image: libvirt: Don't undefine the template machine
Currently, whenever creation of a template is finished, the machine XML
is undefined from libvirt. This is rather counter-productive since if a
user needs to refresh the template, they'd have to re-create the
machine with virt-install; it's much easier to just start the template
machine, refresh it and sanitize it again to create a new template
image. If undefining the template machine is desired, it can still be
done manually.

Signed-off-by: Erik Skultety <eskultet@redhat.com>