make_base_image: Disable the libvirtd service and socket
It may happen that the VM base images were pre-installed with system
libvirt because of some other dependency (e.g. TCK test framework). The
problem is that re-installing libvirt from RPM files directly caused
issues with the default network not to be running before the test suite
was executed. This "hack" disables the system version of libvirt (while
still satisfying the TCK dependency) and allows for a fresh RPM direct
install of libvirt.

Signed-off-by: Erik Skultety <eskultet@redhat.com>