make_base_image: Allow SSH keys to be removed during the sysprep step
The SSH keys were preserved during the sysprep step only because we
relied solely on the Ansible playbooks to upload the "correct" ones and
we didn't want virt-sysprep to remove them immediately. With cloud-init
however, we don't care about any SSH keys anymore, because the right
ones will be uploaded on the first boot.

Signed-off-by: Erik Skultety <eskultet@redhat.com>