Introduce a configmap module
This module defines a config map class which apart from being a
singleton (i.e. a global instance), it also behaves like a dictionary
which is what we'd expect from such an object. The obvious advantage
over a standard dictionary is that we can add custom methods like
loading configuration from disk in the future.

Signed-off-by: Erik Skultety <eskultet@redhat.com>