make_base_image: Don't explicitly disable libvirt services
The original problem was a combination of some of the libvirt
functional test suite test cases which test daemon restart,
prior existing stock installation of libvirt in the environment and
DNF upgrade of the libvirt packages which combined together caused the
default network not to be active. This change should not be needed in
its current shape and form.

Signed-off-by: Erik Skultety <eskultet@redhat.com>