provisioner: libvirt: Assume default storage pool instead of 'base_imgs'
This was some leftover from the original idea where I used a separate
pool for the base images on the worker nodes, but we cannot assume
anything apart from the default storage pool from libvirt. The proper
fix would be to make this configurable, but that's a patch for another
time, let's just provide a simpler fix in the meantime.

Signed-off-by: Erik Skultety <eskultet@redhat.com>