provisioner: machine: Wait for machine readiness by waiting on SSH
Current cloud-init based solution relying on the phone_home module
contacting an ad-hoc HTTP server when machine setup is finished
requires a proper firewall setup so that packets originating in the
libvirt network targeting the host on random ports aren't dropped.
Replace this solution with one that works out-of-the-box on essentially
any machine that was configured with SSH. It does brings some
non-determinism with it, namely a fixed timeout which we assume is
enough for a machine to boot up - this is the tax for convenience and
code simplicity.

Signed-off-by: Erik Skultety <eskultet@redhat.com>