provisioner: machine: Make the SSH key a mandatory argument
It's application's job to figure out what key to use and then pass it
to other interfaces, e.g. Machine in this case. Since the application
module does that already and returns an exception if a suitable key
could not be used, we can make the SSH key argument mandatory here.

Signed-off-by: Erik Skultety <eskultet@redhat.com>