machine: Improve handling of SSH's "Name or service not known"
Depending on Paramiko library version and libvirt dnsmasq's leases file
trying to SSH to a fresh machine during its boot sequence may lead to 2
different types of exception signaling that the network is not up
yet and we have to wait:
- OSError with errno -2 coming from Python's socket library
- NoValidConnectionsError with errno 113 coming from Paramiko itself

Paramiko states in their docs that NoValidConnectionsError is a wrapper
exception type that wraps *all* connection related errors which is not
true when it comes to calling 'getaddrinfo' to resolve a hostname
where they only implement a "fail fast" mechanism and so
socket.gaierror instead of NoValidConnectionsError is raised.
Unfortunately for us this means we have to handle both cases as we
have essentially no control over which one will be thrown at a given
time.

Signed-off-by: Erik Skultety <eskultet@redhat.com>