machine: Increase the SSH wait timeout
Trying to spin up several VMs at the same time occasionally leads to
DHCP (NetworkManager) in the guest being a bit too slow to allocate an
IP address within the current 30s time frame which results in
unnecessary provision failures. Taking a more conservative approach by
increasing the timeout to 60s should give the VM enough time to reserve
and IP address. Note that this is only the worst case scenario and
since we wait in 2s intervals, we'll connect to the machine a lot
quicker than those 60s most of the time.

Signed-off-by: Erik Skultety <eskultet@redhat.com>