machine: provision: Split the combined exception handler in two
The problem with combining paramiko's NoValidConnectionsError and
generic OSError in a single handler as commit 96ca15f5 did is that
NoValidConnectionsError is a subclass of OSError and hence the
EAI_NONAME errno check is performed against it. The problem is that
NoValidConnectionsError is structured differently than a plain OSError
in that it doesn't have a top-level 'errno' attribute. For the reason
above we also cannot define the handler for OSError as the first one
because the 'except' clauses are evaluated in order.

Fixes: 96ca15f5

Signed-off-by: Erik Skultety <eskultet@redhat.com>