machine: Make virt-install re-try the install if it failed
One of the more common failures we could encounter when several
instances of virt-install are trying to create a new VM is that all
of these are trying to refresh virt-install's own 'boot-scratch'
storage pool. That should not be a fatal error and instead virt-install
should re-try to refresh the pool. Until that behaviour is agreed upon
here:
https://github.com/virt-manager/virt-manager/issues/498

we need to provide a workaround ourselves. Direct string matching is
very error prone so this patch gives a grace period of 3 attempts to
provision a VM before resulting in a fatal error.

Signed-off-by: Erik Skultety <eskultet@redhat.com>