application: Drop 'cmdlinestr' from the execution exception's message
The try block is fairly big covering multiple error paths. One of the
possible error paths is that we failed to re-connect to the machine for
some reason which means we don't have 'cmdlinestr' defined at that
point in time. Proper fix would be to define custom exception types
and select behaviour depending on that. That's a task for another day.

Signed-off-by: Erik Skultety <eskultet@redhat.com>